major changes

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -22,7 +22,7 @@ class CreateUsersTable extends Migration
             $table->string('password');
             $table->string('address');
             $table->string('phone');
-            $table->float('balance')->default(0);
+            $table->decimal('balance', 19, 4)->default(0);
             $table->rememberToken();
             $table->timestamps();
         });
